refactor: added types, constants and api in a lib folder
